document coastlines but NOTE
NOTE: I have commented-out the plotting code.  I may try adding it back
later, but it really doesn't save the user from doing much work, given
that the documentation has copy/paste code examples. The problem is with
the ! method, which makes a whole new plot.  I think there is a trick
where it is supposed to find a way to know if a plot already exists.  I
wasted time on this and I think it's only a minor help to the user.