Add Amsr type and read_amsr(): preliminary because...
I'm not sure whether I want to make it read all the data.  That's what
we do in oce, but I'm not sure if making the data slot be a Dict is
such a great idea (memory-wise).

For ad2cp data in oce, we decided let users specify what they want. That
was easier to code for subsequent things like plotting etc.

I'll think about this, given usage patterns.