make objects mutable ...
I've read that this makes things slower, but it is *very* convenient for
users to be able to construct subsets (e.g. downcasts of ctd traces), as
the other alternative is to e.g.

    ctd = (read something)
    metadata = ctd.metadata
    data = ctd.data
    data = data[keep, :]
    ctd_new = Ctd(metadata, data)

However, of course, if I create a subset() that does like the above,
then users will find that convenient (and like R/oce) so I'll keep that
in mind.