Break plot.jl up into one file per function
These functions are liable to get quite long, and breaking things up may
make it easier to work on the code.  (I don't know a good way to narrow
to functions in neovim ... going to the end of the file, which I do *a
lot* skips past the end of the narrowed region. I think emacs did this
better, but there's no way I'm going back to emacs.)